feat(featured): no more change on close
We think it is bad practice to update fields on a close. Mainly because it won't add any extra information, because the feature is closed and the validity of the change is from t_close to t_close, which is void. If the requirement was to update the last record, than a change record with the same validity as the current will fulfil this.